+++ src/infrastructure/implementations/java-tests-runner.go
@@ -17,6 +17,13 @@ import (
 
 type JavaTestsRunner struct{}
 
+var templateArchivePathTemplate = "%s/template.zip"
+var testsArchivePathTemplate = "%s/tests.zip"
+var submissionArchivePathTemplate = "%s/submission.zip"
+
+var defaultErrorLineOnEmptyErrorLines = "[ERROR] We had an error while running the tests. It's possible that the tests execution time exceeded our limit. Please try again or report the issue if it persists."
+
+// SaveArchivesInFS saves the archives needed to run the tests in the file system
 func (javaTestsRunner *JavaTestsRunner) SaveArchivesInFS(dto *dtos.TestArchivesDTO) error {
 	// Ensure the directory doesn't exist
 	path := fmt.Sprintf(
@@ -38,7 +45,7 @@ func (javaTestsRunner *JavaTestsRunner) SaveArchivesInFS(dto *dtos.TestArchivesD
 	// Save the archives
 	err = javaTestsRunner.saveArchiveInFS(
 		dto.LanguageTemplateArchive,
-		fmt.Sprintf("%s/template.zip", path),
+		fmt.Sprintf(templateArchivePathTemplate, path),
 	)
 	if err != nil {
 		return err
@@ -46,7 +53,7 @@ func (javaTestsRunner *JavaTestsRunner) SaveArchivesInFS(dto *dtos.TestArchivesD
 
 	err = javaTestsRunner.saveArchiveInFS(
 		dto.TestsArchive,
-		fmt.Sprintf("%s/tests.zip", path),
+		fmt.Sprintf(testsArchivePathTemplate, path),
 	)
 	if err != nil {
 		return err
@@ -54,7 +61,7 @@ func (javaTestsRunner *JavaTestsRunner) SaveArchivesInFS(dto *dtos.TestArchivesD
 
 	err = javaTestsRunner.saveArchiveInFS(
 		dto.SubmissionArchive,
-		fmt.Sprintf("%s/submission.zip", path),
+		fmt.Sprintf(submissionArchivePathTemplate, path),
 	)
 	if err != nil {
 		return err
@@ -69,6 +76,8 @@ func (javaTestsRunner *JavaTestsRunner) saveArchiveInFS(fileBytes *[]byte, path 
 	return err
 }
 
+// MergeArchives merges the content of the template archive, teacher's tests archive and student's
+// submission archive into a single directory that will be used to run the tests
 func (javaTestsRunner *JavaTestsRunner) MergeArchives(submissionUUID string) error {
 	// Unzip the archives
 	err := javaTestsRunner.unzipArchives(submissionUUID)
@@ -114,7 +123,7 @@ func (javaTestsRunner *JavaTestsRunner) unzipArchives(submissionUUID string) err
 	archivesManager := ArchivesManagerImplementation{}
 
 	err := archivesManager.ExtractArchive(
-		fmt.Sprintf("%s/template.zip", submissionPathPrefix),
+		fmt.Sprintf(templateArchivePathTemplate, submissionPathPrefix),
 		fmt.Sprintf("%s/template", submissionPathPrefix),
 	)
 	if err != nil {
@@ -123,7 +132,7 @@ func (javaTestsRunner *JavaTestsRunner) unzipArchives(submissionUUID string) err
 	}
 
 	err = archivesManager.ExtractArchive(
-		fmt.Sprintf("%s/tests.zip", submissionPathPrefix),
+		fmt.Sprintf(testsArchivePathTemplate, submissionPathPrefix),
 		fmt.Sprintf("%s/tests", submissionPathPrefix),
 	)
 	if err != nil {
@@ -132,7 +141,7 @@ func (javaTestsRunner *JavaTestsRunner) unzipArchives(submissionUUID string) err
 	}
 
 	err = archivesManager.ExtractArchive(
-		fmt.Sprintf("%s/submission.zip", submissionPathPrefix),
+		fmt.Sprintf(submissionArchivePathTemplate, submissionPathPrefix),
 		fmt.Sprintf("%s/submission", submissionPathPrefix),
 	)
 	if err != nil {
@@ -153,7 +162,7 @@ func (javaTestsRunner *JavaTestsRunner) deleteArchives(submissionUUID string) er
 	archivesManager := ArchivesManagerImplementation{}
 
 	err := archivesManager.DeleteArchive(
-		fmt.Sprintf("%s/template.zip", submissionPathPrefix),
+		fmt.Sprintf(templateArchivePathTemplate, submissionPathPrefix),
 	)
 	if err != nil {
 		log.Println("Error while deleting template archive", err)
@@ -161,7 +170,7 @@ func (javaTestsRunner *JavaTestsRunner) deleteArchives(submissionUUID string) er
 	}
 
 	err = archivesManager.DeleteArchive(
-		fmt.Sprintf("%s/tests.zip", submissionPathPrefix),
+		fmt.Sprintf(testsArchivePathTemplate, submissionPathPrefix),
 	)
 	if err != nil {
 		log.Println("Error while deleting tests archive", err)
@@ -169,7 +178,7 @@ func (javaTestsRunner *JavaTestsRunner) deleteArchives(submissionUUID string) er
 	}
 
 	err = archivesManager.DeleteArchive(
-		fmt.Sprintf("%s/submission.zip", submissionPathPrefix),
+		fmt.Sprintf(submissionArchivePathTemplate, submissionPathPrefix),
 	)
 	if err != nil {
 		log.Println("Error while deleting submission archive", err)
@@ -179,8 +188,15 @@ func (javaTestsRunner *JavaTestsRunner) deleteArchives(submissionUUID string) er
 	return nil
 }
 
+// RunTests runs the tests and returns the result
 func (javaTestsRunner *JavaTestsRunner) RunTests(submissionUUID string) (dto *dtos.TestResultDTO, err error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Minute)
+	timeLimitInMinutes := 2.0
+	timeLimitInSeconds := int(timeLimitInMinutes * 60)
+
+	ctx, cancel := context.WithTimeout(
+		context.Background(),
+		time.Duration(timeLimitInSeconds)*time.Second,
+	)
 	defer cancel()
 
 	// Delete the submission directory at the end
@@ -195,8 +211,9 @@ func (javaTestsRunner *JavaTestsRunner) RunTests(submissionUUID string) (dto *dt
 
 	// Prepare the command
 	testAndBuildCommand := fmt.Sprintf(
-		"cd %s && timeout 1m mvn clean test",
+		"cd %s && timeout %dm mvn clean test",
 		submissionPath,
+		timeLimitInSeconds,
 	)
 
 	cmd := exec.CommandContext(
@@ -214,6 +231,10 @@ func (javaTestsRunner *JavaTestsRunner) RunTests(submissionUUID string) (dto *dt
 		errorLines := javaTestsRunner.getErrorLinesFromOutput(string(out))
 		errorLines = javaTestsRunner.sanitizeConsoleTextLines(errorLines)
 
+		if len(errorLines) == 0 {
+			errorLines = []string{defaultErrorLineOnEmptyErrorLines}
+		}
+
 		return &dtos.TestResultDTO{
 			SubmissionUUID: submissionUUID,
 			TestsPassed:    false,
@@ -222,7 +243,7 @@ func (javaTestsRunner *JavaTestsRunner) RunTests(submissionUUID string) (dto *dt
 	}
 
 	// Parse success lines
-	successLines := javaTestsRunner.getSuccessLinesFromOutput(string(out))
+	successLines := javaTestsRunner.getResultLinesFromSuccessOutput(string(out))
 	successLines = javaTestsRunner.sanitizeConsoleTextLines(successLines)
 
 	return &dtos.TestResultDTO{
@@ -232,19 +253,28 @@ func (javaTestsRunner *JavaTestsRunner) RunTests(submissionUUID string) (dto *dt
 	}, nil
 }
 
+// getErrorLinesFromOutput returns the lines starting with the `[ERROR]` prefix from the output
 func (javaTestsRunner *JavaTestsRunner) getErrorLinesFromOutput(output string) []string {
 	errorRegex := regexp.MustCompile(`(?m)^\[ERROR\].*$`)
 	errorLines := errorRegex.FindAllString(output, -1)
 	return errorLines
 }
 
+// sanitizeConsoleTextLines removes the lines that are not relevant for the user or can contain
+// sensitive information
 func (javaTestsRunner *JavaTestsRunner) sanitizeConsoleTextLines(textLines []string) []string {
 	sanitizedTextLines := []string{}
 
 	regExpToReplace := []dtos.ReplaceRegexDTO{
+		// Remove possible path to the tests execution directory
 		{
 			Regexp:      *regexp.MustCompile(`\/[a-zA-Z0-9_\-]+(?:\/[a-zA-Z0-9_\-]+)*\/template`),
 			Replacement: "****/****",
+		},
+		// Remove lines starting with [WARNING]
+		{
+			Regexp:      *regexp.MustCompile(`(?m)^\[WARNING\].*$`),
+			Replacement: "",
 		},
 	}
 
@@ -253,13 +283,17 @@ func (javaTestsRunner *JavaTestsRunner) sanitizeConsoleTextLines(textLines []str
 			errorLine = regExp.Regexp.ReplaceAllString(errorLine, regExp.Replacement)
 		}
 
-		sanitizedTextLines = append(sanitizedTextLines, errorLine)
+		if len(errorLine) > 0 {
+			sanitizedTextLines = append(sanitizedTextLines, errorLine)
+		}
 	}
 
 	return sanitizedTextLines
 }
 
-func (javaTestsRunner *JavaTestsRunner) getSuccessLinesFromOutput(output string) []string {
+// getResultLinesFromSuccessOutput returns the lines of the output starting from the tests results
+// summary / header line until the end
+func (javaTestsRunner *JavaTestsRunner) getResultLinesFromSuccessOutput(output string) []string {
 	// Header line
 	successRegex := regexp.MustCompile(`\[INFO\] Tests run: \d+, Failures: \d+, Errors: \d+, Skipped: \d+`)